`ProcInfo` has been moved to its own Header-File-and-Translation-Unit.
