+++ Makefile
@@ -1,6 +1,7 @@
 oomstaller : oomstaller.cpp
 	$(CXX) -pthread -std='gnu++17' -Os -Wall $(CXXFLAGS) -o 'oomstaller' \
 		'readMemInfo.cpp' \
+		'ProcInfo.cpp' \
 		'oomstaller.cpp'
 
 clean :
